refactor: move dto package
